in progress

+++ app/src/main/java/com/bykea/pk/partner/models/response/BatchBookingDropoff.java
@@ -6,24 +6,24 @@ import com.google.gson.annotations.SerializedName;
 public class BatchBookingDropoff {
     @SerializedName("lat")
     @Expose
-    private Integer lat;
+    private double lat;
     @SerializedName("lng")
     @Expose
-    private Integer lng;
+    private double lng;
 
-    public Integer getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(Integer lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public Integer getLng() {
+    public double getLng() {
         return lng;
     }
 
-    public void setLng(Integer lng) {
+    public void setLng(double lng) {
         this.lng = lng;
     }
 }
